combined blink processor
